+++ SieuThiMini/src/DTO/ChiTietPhieuNhapHangDTO.java
@@ -3,14 +3,14 @@ package DTO;
 public class ChiTietPhieuNhapHangDTO {
     private int maphieunhaphang;
     private int masanpham;
-    private double dongia;
+    private long dongia;
     private int soluong;
-    private double thanhtien;
+    private long thanhtien;
 
     public ChiTietPhieuNhapHangDTO() {
     }
 
-    public ChiTietPhieuNhapHangDTO(int maphieunhaphang, int masanpham, double dongia, int soluong, double thanhtien) {
+    public ChiTietPhieuNhapHangDTO(int maphieunhaphang, int masanpham, long dongia, int soluong, long thanhtien) {
         this.maphieunhaphang = maphieunhaphang;
         this.masanpham = masanpham;
         this.dongia = dongia;
@@ -22,10 +22,10 @@ public class ChiTietPhieuNhapHangDTO {
     public void setMaPNH(int maphieunhaphang) { this.maphieunhaphang = maphieunhaphang; }
     public int getMaSP() { return masanpham; }
     public void setMaSP(int masanpham) { this.masanpham = masanpham; }
-    public double getDonGia() { return dongia; }
-    public void setDonGia(double dongia) { this.dongia = dongia; }
+    public long getDonGia() { return dongia; }
+    public void setDonGia(long dongia) { this.dongia = dongia; }
     public int getSoLuong() { return soluong; }
     public void setSoLuong(int soluong) { this.soluong = soluong; }
-    public double getThanhTien() {return this.soluong * this.dongia;}
-    public void setThanhTien(double thanhtien) { this.thanhtien = thanhtien; }
+    public long getThanhTien() {return this.soluong * this.dongia;}
+    public void setThanhTien(long thanhtien) { this.thanhtien = thanhtien; }
 }
