+++ SieuThiMini/src/DTO/ChuongTrinhKhuyenMaiSpDTO.java
@@ -3,12 +3,12 @@ package DTO;
 public class ChuongTrinhKhuyenMaiSpDTO {
     private int chuongTrinhKhuyenMaiId;
     private int sanPhamId;
-    private int giaTriGiam;
+    private long giaTriGiam;
 
     public ChuongTrinhKhuyenMaiSpDTO() {
     }
 
-    public ChuongTrinhKhuyenMaiSpDTO(int chuongTrinhKhuyenMaiId, int sanPhamId, int giaTriGiam) {
+    public ChuongTrinhKhuyenMaiSpDTO(int chuongTrinhKhuyenMaiId, int sanPhamId, long giaTriGiam) {
         this.chuongTrinhKhuyenMaiId = chuongTrinhKhuyenMaiId;
         this.sanPhamId = sanPhamId;
         this.giaTriGiam = giaTriGiam;
@@ -18,6 +18,6 @@ public class ChuongTrinhKhuyenMaiSpDTO {
     public void setChuongTrinhKhuyenMaiId(int chuongTrinhKhuyenMaiId) { this.chuongTrinhKhuyenMaiId = chuongTrinhKhuyenMaiId; }
     public int getSanPhamId() { return sanPhamId; }
     public void setSanPhamId(int sanPhamId) { this.sanPhamId = sanPhamId; }
-    public int getGiaTriGiam() { return giaTriGiam; }
-    public void setGiaTriGiam(int giaTriGiam) { this.giaTriGiam = giaTriGiam; }
+    public long getGiaTriGiam() { return giaTriGiam; }
+    public void setGiaTriGiam(long giaTriGiam) { this.giaTriGiam = giaTriGiam; }
 }
