+++ SieuThiMini/src/DTO/SanPhamDTO.java
@@ -4,15 +4,15 @@ public class SanPhamDTO {
     private  int masanpham;
     private  String tensanpham;
     private int soluong;
-    private int dongia;
+    private long dongia;
     private String donvitinh;
     private int maLoai;
     private int maHang;
-    private double giaNhap;
+    private long giaNhap;
     private double loiNhuan;
     public SanPhamDTO(){}
 
-    public SanPhamDTO(int masanpham, String tensanpham, int soluong, int dongia,double giaNhap,double loiNhuan, String donvitinh,int maHang,int maLoai){
+    public SanPhamDTO(int masanpham, String tensanpham, int soluong, long dongia,long giaNhap,double loiNhuan, String donvitinh,int maHang,int maLoai){
         this.masanpham = masanpham;
         this.tensanpham = tensanpham;
         this.soluong = soluong;
@@ -30,16 +30,16 @@ public class SanPhamDTO {
     public void setTensanpham(String tensanpham) {this.tensanpham = tensanpham;}
     public int getSoluong() {return soluong;}
     public void setSoluong(int soluong) {this.soluong = soluong;}
-    public int getDongia() {return dongia;}
-    public void setDongia(int dongia) {this.dongia = dongia;}
+    public long getDongia() {return dongia;}
+    public void setDongia(long dongia) {this.dongia = dongia;}
     public String getDonvitinh() {return donvitinh;}
     public void setDonvitinh(String donvitinh) {this.donvitinh = donvitinh;}
     public int getMaLoai() { return maLoai; }
     public void setMaLoai(int maLoai) { this.maLoai = maLoai; }
     public int getMaHang() { return maHang; }
     public void setMaHang(int maHang) { this.maHang = maHang; }
-    public double getGiaNhap() { return giaNhap; }
-    public void setGiaNhap(double giaNhap) { this.giaNhap = giaNhap; }
+    public long getGiaNhap() { return giaNhap; }
+    public void setGiaNhap(long giaNhap) { this.giaNhap = giaNhap; }
 
     public double getLoiNhuan() { return loiNhuan; }
     public void setLoiNhuan(double loiNhuan) { this.loiNhuan = loiNhuan; }
